% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyShip.R
\name{classify}
\alias{classify}
\title{Define the ship type given a ship code.}
\usage{
classify(shipcode)
}
\arguments{
\item{shipcode}{a code provided in the AIS data}
}
\value{
Return a ship type, which can be Cargo, Fishing, Military, Not_available, Passenger, Pleasure, Tanker, Tugboat, and Others
}
\description{
Define the ship type given a ship code.
}
\examples{
classify(70)
classify(shipcode = 70)
}
