% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperatingTime.R
\name{operatingtime}
\alias{operatingtime}
\title{Calculate the total operating time for each vessel and each ship type}
\usage{
operatingtime(
  AISdata,
  activityspeed = 3,
  maxLON = -90,
  minLON = -110,
  maxLAT = 30,
  minLAT = 20
)
}
\arguments{
\item{AISdata}{Automatic Identification System (AIS) data downloaded from the U.S. Coast Guard (https://marinecadastre.gov/ais/). You can use the command: AISdata <- read.csv("AISdataname.csv", header = T) and input AISdata in the function}

\item{activityspeed}{Definition of the operating speed. Default speed is 3 knot, meaning that only activity segment with over 3 knots will be considered.}

\item{maxLON}{Maximum longitude of the study area}

\item{minLON}{Minimum longitude of the study area}

\item{maxLAT}{Maximum latitude of the study area}

\item{minLAT}{Minimum latitude of the study area}
}
\value{
A list of two data frames. The first data frame is the operating time for each vessel (MMSI); the second data frame is operating time for ship types.
}
\description{
Calculate the total operating time for each vessel and each ship type
}
\examples{
operatingtime(AISdata)
operatingtime(AISdata, activityspeed = 5)
operatingtime(AISdata, activityspeed = 5 , maxLON = -90, minLON = -110, maxLAT = 30, minLAT = 20)
}
