% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotVessel.R
\name{plotvessel}
\alias{plotvessel}
\title{Plot vessel locations}
\usage{
plotvessel(
  AISdata,
  studyship = "Cargo",
  maxLON = -94,
  minLON = -96,
  maxLAT = 30,
  minLAT = 27
)
}
\arguments{
\item{AISdata}{Automatic Identification System (AIS) data downloaded from the U.S. Coast Guard (https://marinecadastre.gov/ais/). You can use the command: AISdata <- read.csv("AISdataname.csv", header = T) and input AISdata in the function}

\item{studyship}{The observed ship type; default is Cargo. Other types include Fishing, Military, Not_available, Passenger, Pleasure, Tanker, Tugboat, Others.}

\item{maxLON}{Maximum longitude of the study area}

\item{minLON}{Minimum longitude of the study area}

\item{maxLAT}{Maximum latitude of the study area}

\item{minLAT}{Minimum latitude of the study area}
}
\value{
Map of signal locations (Viewer Panel)
}
\description{
Plot vessel locations
}
\examples{
plotvessel(AISdata)
plotvessel(AISdata, 'Fishing')
plotvessel(AISdata, studyship = 'Fishing', maxLON = -94, minLON = -96, maxLAT = 30, minLAT = 27)
}
