% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeedDistribution.R
\name{speeddist}
\alias{speeddist}
\title{Plot speed histogram for a ship type}
\usage{
speeddist(
  AISdata,
  studyship = "Cargo",
  maxLON = -94,
  minLON = -96,
  maxLAT = 30,
  minLAT = 27
)
}
\arguments{
\item{AISdata}{Automatic Identification System (AIS) data downloaded from the U.S. Coast Guard (https://marinecadastre.gov/ais/). You can use the command: AISdata <- read.csv("AISdataname.csv", header = T) and input AISdata in the function}

\item{studyship}{The observed ship type; default is Cargo. Other types include Fishing, Military, Not_available, Passenger, Pleasure, Tanker, Tugboat, Others}

\item{maxLON}{Maximum longitude of the study area}

\item{minLON}{Minimum longitude of the study area}

\item{maxLAT}{Maximum latitude of the study area}

\item{minLAT}{Minimum latitude of the study area}
}
\value{
Histogram of the speed. An object of class "histogram" is a list with components. Please refer to hist
}
\description{
Plot speed histogram for a ship type
}
\examples{
speeddist(AISdata)
speeddist(AISdata, 'Fishing')
speeddist(AISdata, studyship = 'Fishing', maxLON = -94, minLON = -96, maxLAT = 30, minLAT = 27)
}
