% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShipSummaryStat.R
\name{summarizeship}
\alias{summarizeship}
\title{Summarize the number and dimensions for each shiptype within study areas}
\usage{
summarizeship(
  AISdata = AISdata,
  maxLON = -90,
  minLON = -110,
  maxLAT = 30,
  minLAT = 20
)
}
\arguments{
\item{AISdata}{Automatic Identification System (AIS) data downloaded from the U.S. Coast Guard (https://marinecadastre.gov/ais/). You can use the command: AISdata <- read.csv("AISdataname.csv", header = T) and input AISdata in the function}

\item{maxLON}{Maximum longitude of the study area}

\item{minLON}{Minimum longitude of the study area}

\item{maxLAT}{Maximum latitude of the study area}

\item{minLAT}{Minimum latitude of the study area}
}
\value{
A data frame with the number and dimension of ships for each ship type
}
\description{
Summarize the number and dimensions for each shiptype within study areas
}
\examples{
summarizeship(AISdata, maxLON=-80, minLON=-100, maxLAT= 40, minLAT= 20)
summarizeship(AISdata)
}
